refactoring

+++ scripts/clear.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 # filename: clear.sh
 
+if [ "$1" == "model" ];then
+    rm model_export/*
+fi
 
-rm datasets/*rec datasets/*idx datasets/*lst
+if [ "$1" == "data" ];then
+    rm datasets/*rec datasets/*idx datasets/*lst
+fi
